+++ js/background/main.js
@@ -1,4 +1,4 @@
 import { initializeServiceWorker } from './background.js';
 
-console.log(`Spaces ${chrome.runtime.getManifest().version}`)
+console.log(`Spaces ${chrome.runtime.getManifest().version}`);
 initializeServiceWorker();